# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Budgetarian API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /warranties:
    get:
      operationId: listWarranties
      tags: [warranties]
      summary: List warranties
      responses:
        "200":
          description: List of warranties
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListWarrantiesResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      operationId: createWarranty
      tags: [warranties]
      summary: Create a warranty entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWarrantyBody"
      responses:
        "201":
          description: Warranty created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Warranty"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /warranties/expiring-soon:
    get:
      operationId: getExpiringSoonWarranties
      tags: [warranties]
      summary: Get warranties expiring within 30 days
      responses:
        "200":
          description: Warranties expiring soon
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListWarrantiesResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /warranties/{id}:
    patch:
      operationId: updateWarranty
      tags: [warranties]
      summary: Update a warranty entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWarrantyBody"
      responses:
        "200":
          description: Updated warranty
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Warranty"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: deleteWarranty
      tags: [warranties]
      summary: Delete a warranty entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # === TIPS ===
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # === DASHBOARD ===
...
    Warranty:
      type: object
      required: [id, userId, productName, purchaseDate, expiryDate, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        productName:
          type: string
        store:
          type: string
          nullable: true
        purchaseDate:
          type: string
          format: date
        expiryDate:
          type: string
          format: date
        receiptPath:
          type: string
          nullable: true
          description: Object path for the receipt image stored in object storage
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    ListWarrantiesResponse:
      type: object
      required: [warranties]
      properties:
        warranties:
          type: array
          items:
            $ref: "#/components/schemas/Warranty"
    CreateWarrantyBody:
      type: object
      required: [productName, purchaseDate, expiryDate]
      properties:
        productName:
          type: string
          minLength: 1
        store:
          type: string
        purchaseDate:
          type: string
          format: date
        expiryDate:
          type: string
          format: date
        receiptPath:
          type: string
          description: Object path returned from object storage upload
        notes:
          type: string
    UpdateWarrantyBody:
      type: object
      properties:
        productName:
          type: string
        store:
          type: string
        purchaseDate:
          type: string
          format: date
        expiryDate:
          type: string
          format: date
        receiptPath:
          type: string
        notes:
          type: string

    # === TIPS ===
...