# lib/api-zod/src/generated/types/warranty.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Budgetarian API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Warranty {
  id: number;
  userId: string;
  productName: string;
  store?: string | null;
  purchaseDate: Date;
  expiryDate: Date;
  /** Object path for the receipt image stored in object storage */
  receiptPath?: string | null;
  notes?: string | null;
  createdAt: Date;
}